refactor(map): extract Pass 6b.5 noise perturbation to its own part file (Refs #2683)
Pass 6b.5 noise perturbation pushed
packages/colonizethis_map/lib/src/tile_map_generator_terrain_assign.dart
to 1032 physical lines, exceeding the 1000-line repo.part_unit_size
ceiling. Move _applyTerrainNoisePerturbation, _perturbOneBlobWithNoise,
and _smoothNoiseAt into a new part file
tile_map_generator_terrain_noise.dart as a library-private extension on
_TileMapGenTerrainResource. Behavior is unchanged; the call site in
_assignNonMountainInComponent still resolves to the same methods via the
extension. terrain_assign drops to 935 lines.
